Extracted sdk_examples library from readme & validation
