+++ .tools/sdk_examples/sdk_examples.egg-info/PKG-INFO
@@ -1,0 +1,41 @@
+Metadata-Version: 2.1
+Name: sdk-examples
+Version: 0.1.0
+Summary: Extract SDK Code Examples Data
+Home-page: https://github.com/awsdocs/aws-doc-sdk-examples
+Author: David Souther
+Author-email: dpsouth@amazon.com
+Project-URL: Bug Reports, https://github.com/awsdocs/aws-doc-sdk-examples/issues
+Keywords: aws docs,aws code examples
+Classifier: Development Status :: 4 - Beta
+Classifier: Intended Audience :: Developers
+Classifier: Topic :: Software Development :: Build Tools
+Classifier: License :: OSI Approved :: Apache 2 License
+Classifier: Programming Language :: Python :: 3
+Classifier: Programming Language :: Python :: 3.7
+Classifier: Programming Language :: Python :: 3.8
+Classifier: Programming Language :: Python :: 3.9
+Classifier: Programming Language :: Python :: 3.10
+Classifier: Programming Language :: Python :: 3.11
+Classifier: Programming Language :: Python :: 3 :: Only
+Requires-Python: >=3.7, <4
+Description-Content-Type: text/markdown
+
+# AWS SDK Code Examples Extractor
+
+This library allows AWS SDK Code Examples engineers programmatic access to the range of metadata in SoS.
+
+## Motivation:
+
+1. SoS Metadata
+1. Snippet Extraction
+1. WRITEME
+1. Ailly
+1. Zexii
+1. Directories / Developer Center
+
+all have some notion of "do something on S service in K language/SDK"
+
+`sdk_examples` provides a convenient entry point, `load`, which will review a folder structure for `.doc_gen` SoS metadata and embedded SOS snippet sections.
+
+This can then be used further for validation, reviewing snippets, etc.
