+++ .tools/sdk_examples/sdk_examples.egg-info/SOURCES.txt
@@ -1,0 +1,20 @@
+README.md
+pyproject.toml
+setup.py
+sdk_examples/__init__.py
+sdk_examples/config.py
+sdk_examples/doc_gen.py
+sdk_examples/file_utils.py
+sdk_examples/metadata.py
+sdk_examples/metadata_errors.py
+sdk_examples/metadata_validator.py
+sdk_examples/sdks.py
+sdk_examples/services.py
+sdk_examples/snippets.py
+sdk_examples.egg-info/PKG-INFO
+sdk_examples.egg-info/SOURCES.txt
+sdk_examples.egg-info/dependency_links.txt
+sdk_examples.egg-info/top_level.txt
+tests/test_metadata.py
+tests/test_sdks.py
+tests/test_services.py
