+++ .tools/sdk_examples/sdk_examples.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+sdk_examples
