+++ .tools/sdk_examples/sdk_examples/sdks.py
@@ -2,8 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from typing import Self, Optional
-import metadata_errors
-from metadata_errors import MetadataErrors, MetadataParseError, check_mapping
+from sdk_examples import metadata_errors
+from sdk_examples.metadata_errors import (
+    MetadataErrors,
+    MetadataParseError,
+    check_mapping,
+)
 from dataclasses import dataclass, field
 
 
@@ -19,20 +23,23 @@ class SdkApiRef:
     name: str
     link_template: Optional[str]
 
-    def from_yaml(yaml: dict[str, str] | None, errors: MetadataErrors) -> Self:
+    def from_yaml(yaml: dict[str, str] | None, errors: MetadataErrors) -> Self | None:
         if yaml is None:
             return None
         uid = yaml.get("uid")
         name = check_mapping(yaml.get("name"), "api_ref.name")
         link_template = yaml.get("link_template")
 
+        e = len(errors)
+
         if not uid:
             errors.append(metadata_errors.MissingField(field="api_ref.uid"))
         if isinstance(name, MetadataParseError):
             errors.append(name)
-            name = ""
 
-        return SdkApiRef(uid, name, link_template)
+        if e == len(errors):
+            return SdkApiRef(uid, name, link_template)
+        return None
 
 
 @dataclass
@@ -54,7 +61,7 @@ class SdkVersion:
     title_override: Optional[SdkTitleOverride] = field(default=None)
 
     @staticmethod
-    def from_yaml(version: int, yaml: dict[str, any]) -> (Self, MetadataErrors):
+    def from_yaml(version: int, yaml: dict[str, any]) -> Self | MetadataErrors:
         errors = MetadataErrors()
         long = check_mapping(yaml.get("long"), "long")
         short = check_mapping(yaml.get("short"), "short")
@@ -88,25 +95,23 @@ class SdkVersion:
 
         if isinstance(long, MetadataParseError):
             errors.append(long)
-            long = ""
         if isinstance(short, MetadataParseError):
             errors.append(short)
-            short = ""
         api_ref = SdkApiRef.from_yaml(yaml.get("api_ref"), errors)
 
-        return (
-            SdkVersion(
-                version=version,
-                long=long,
-                short=short,
-                expanded=expanded,
-                guide=guide,
-                api_ref=api_ref,
-                caveat=caveat,
-                bookmark=bookmark,
-                title_override=title_override,
-            ),
-            errors,
+        if len(errors) > 0:
+            return errors
+
+        return SdkVersion(
+            version=version,
+            long=long,
+            short=short,
+            expanded=expanded,
+            guide=guide,
+            api_ref=api_ref,
+            caveat=caveat,
+            bookmark=bookmark,
+            title_override=title_override,
         )
 
 
@@ -118,13 +123,12 @@ class Sdk:
     property: str
 
     @staticmethod
-    def from_yaml(name: str, yaml: dict[str, any]) -> (Self, MetadataErrors):
+    def from_yaml(name: str, yaml: dict[str, any]) -> Self | MetadataErrors:
         errors = MetadataErrors()
         property = yaml.get("property")
         guide = check_mapping(yaml.get("guide"), "guide")
         if isinstance(guide, MetadataParseError):
             errors.append(guide)
-            guide = ""
 
         versions = []
         sdk_versions = yaml.get("sdk", {})
@@ -137,10 +141,13 @@ class Sdk:
             else:
                 versions.append(sdk_version)
 
-        return Sdk(name=name, versions=versions, guide=guide, property=property), errors
+        if len(errors) > 0:
+            return errors
+
+        return Sdk(name=name, versions=versions, guide=guide, property=property)
 
 
-def parse(file: str, yaml: dict[str, any]) -> (dict[str, Sdk], MetadataErrors):
+def parse(file: str, yaml: dict[str, any]) -> dict[str, Sdk]:
     sdks = {}
     errors = MetadataErrors()
 
@@ -154,7 +161,7 @@ def parse(file: str, yaml: dict[str, any]) -> (dict[str, Sdk], MetadataErrors):
                 error.id = name
             errors.extend(sdk)
 
-    return sdks, errors
+    return sdks if len(errors) == 0 else errors
 
 
 if __name__ == "__main__":
