+++ .tools/sdk_examples/tests/test_resources/awsentity_metadata.yaml
