+++ .tools/sdk_examples/tests/test_resources/curated/sources.yaml
