+++ .tools/sdk_examples/tests/test_resources/empty_metadata.yaml
