+++ .tools/sdk_examples/tests/test_resources/empty_services.yaml
