+++ .tools/sdk_examples/tests/test_resources/malformed_metadata.yaml
