+++ .tools/sdk_examples/tests/test_resources/sdks.yaml
