+++ .tools/sdk_examples/tests/test_resources/services.yaml
