+++ .tools/sdk_examples/tests/test_resources/valid_metadata.yaml
