--- .tools/validation/doc_gen.py
@@ -1,38 +1,0 @@
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-import yaml
-
-from typing import Self
-from dataclasses import dataclass, field
-from pathlib import Path
-
-from metadata_errors import MetadataErrors
-from sdks import Sdk, parse as parse_sdks
-from services import Service, parse as parse_services
-from snippets import Snippet
-
-
-@dataclass
-class DocGen:
-    sdks: dict[str, Sdk] = field(default_factory=dict)
-    services: dict[str, Service] = field(default_factory=dict)
-    snippets: dict[str, Snippet] = field(default_factory=dict)
-
-    @staticmethod
-    def from_root(root: Path) -> (Self, MetadataErrors):
-        errors = MetadataErrors()
-
-        with open(root / "sdks.yaml", encoding="utf-8") as file:
-            meta = yaml.safe_load(file)
-            parsed = parse_sdks("sdks.yaml", meta)
-            sdks = errors.maybe_extend(parsed)
-
-        with open(root / "services.yaml", encoding="utf-8") as file:
-            meta = yaml.safe_load(file)
-            services, service_errors = parse_services("services.yaml", meta)
-            errors.extend(service_errors)
-
-        snippets = {}
-
-        return DocGen(sdks=sdks, services=services, snippets=snippets), errors
